sdk: silo cleanup - extra types to separate file
