+++ projects/sdk/src/index.ts
@@ -12,7 +12,7 @@ export { DecimalBigNumber } from "src/classes/DecimalBigNumber";
 
 // Modules
 export { FarmWorkflow, FarmFromMode, FarmToMode } from "src/lib/farm";
-export type { TokenSiloBalance } from "src/lib/silo";
+export type { TokenSiloBalance } from "src/lib/silo/types";
 export type { TokenBalance } from "src/lib/tokens";
 export { AdvancedPipeWorkflow, Clipboard } from "src/lib/depot";
 export type { PipeCallStruct as PipeStruct, AdvancedPipeCallStruct as AdvancedPipeStruct } from "src/lib/depot";
