+++ projects/sdk/src/lib/silo.ts
@@ -1,6 +1,5 @@
 import { ethers, BigNumber, ContractTransaction } from "ethers";
 import { ERC20Token, Token } from "src/classes/Token";
-import { StringMap } from "src/types";
 import { BeanstalkSDK, DataSource } from "./BeanstalkSDK";
 import EventProcessor from "src/lib/events/processor";
 import { EIP712Domain, EIP712TypedData, Permit } from "./permit";
@@ -9,7 +8,7 @@ import {
   DepositTokenPermitMessage,
   DepositTokensPermitMessage,
   sortCratesBySeason,
-  _parseWithdrawalCrates
+  parseWithdrawalCrates
 } from "./silo/utils";
 import { TokenValue } from "src/classes/TokenValue";
 import { MAX_UINT256 } from "src/constants";
@@ -19,66 +18,7 @@ import { DepositOperation } from "./silo/DepositOperation";
 import { Withdraw } from "./silo/Withdraw";
 import { Claim } from "./silo/Claim";
 import { FarmToMode } from "./farm";
-
-/**
- * A Crate is an `amount` of a token Deposited or
- * Withdrawn during a given `season`.
- */
-type BigNumbers = TokenValue;
-export type Crate<T extends BigNumbers = TokenValue> = {
-  /** The amount of this Crate that was created, denominated in the underlying Token. */
-  amount: T;
-  /** The Season that the Crate was created. */
-  season: BigNumber;
-};
-
-/**
- * A "Deposit" represents an amount of a Whitelisted Silo Token
- * that has been added to the Silo.
- */
-export type DepositCrate<T extends BigNumbers = TokenValue> = Crate<T> & {
-  /** The BDV of the Deposit, determined upon Deposit. */
-  bdv: T;
-  /** The total amount of Stalk granted for this Deposit. */
-  stalk: T;
-  /** The Stalk associated with the BDV of the Deposit. */
-  baseStalk: T;
-  /** The Stalk grown since the time of Deposit. */
-  grownStalk: T;
-  /** The amount of Seeds granted for this Deposit. */
-  seeds: T;
-};
-
-export type WithdrawalCrate<T extends BigNumbers = TokenValue> = Crate<T> & {};
-
-/**
- * A "Silo Balance" provides all information
- * about a Farmer's ownership of a Whitelisted Silo Token.
- */
-export type TokenSiloBalance = {
-  deposited: {
-    /** The total amount of this Token currently in the Deposited state. */
-    amount: TokenValue;
-    /** The BDV of this Token currently in the Deposited state. */
-    bdv: TokenValue;
-    /** All Deposit crates. */
-    crates: DepositCrate<TokenValue>[];
-  };
-  withdrawn: {
-    /** The total amount of this Token currently in the Withdrawn state. */
-    amount: TokenValue;
-    /** All Withdrawal crates. */
-    crates: WithdrawalCrate<TokenValue>[];
-  };
-  claimable: {
-    /** The total amount of this Token currently in the Claimable state. */
-    amount: TokenValue;
-    /** All Claimable crates. */
-    crates: Crate<TokenValue>[];
-  };
-};
-
-export type UpdateFarmerSiloBalancesPayload = StringMap<Partial<TokenSiloBalance>>;
+import { DepositCrate, WithdrawalCrate, Crate, TokenSiloBalance } from "./silo/types";
 
 export class Silo {
   static sdk: BeanstalkSDK;
@@ -236,7 +176,7 @@ export class Silo {
 
   //////////////////////// BALANCES ////////////////////////
 
-  private _parseWithdrawalCrates = _parseWithdrawalCrates;
+  private _parseWithdrawalCrates = parseWithdrawalCrates;
 
   private _makeTokenSiloBalance(): TokenSiloBalance {
     return {
