+++ projects/sdk/src/utils/TestUtils/BlockchainUtils.ts
@@ -1,7 +1,7 @@
 import { ethers } from "ethers";
 import { ERC20Token } from "src/classes/Token";
 import { BeanstalkSDK, DataSource } from "src/lib/BeanstalkSDK";
-import { TokenSiloBalance } from "src/lib/silo";
+import { TokenSiloBalance } from "src/lib/silo/types";
 import { TokenValue } from "src/TokenValue";
 import * as addr from "./addresses";
 import { logSiloBalance } from "./log";
